+++ src/resources/routes/admin.php
@@ -19,6 +19,7 @@ Route::resource('master_product', 'MasterProductController')->except(['index']);
 Route::resource('property', 'PropertyController');
 Route::resource('order', 'OrderController');
 Route::resource('media', 'MediaController')->only(['update', 'destroy', 'store']);
+Route::resource('carrier', 'CarrierController');
 Route::resource('payment-method', 'PaymentMethodController')
      ->parameters(['payment-method' => 'paymentMethod']);
 Route::resource('shipping-method', 'ShippingMethodController')