+++ Changelog.md
@@ -1,7 +1,7 @@
 # Vanilo Admin Changelog
 
-## Unreleased
-##### 2023-XX-YY
+## 3.4.0
+##### 2023-01-25
 
 - Added Carrier CRUD
 - Added Shipping method CRUD